% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{center_vars}
\alias{center_vars}
\title{Centre variables}
\usage{
center_vars(
  frame,
  choose,
  center = TRUE,
  scale = FALSE,
  ...,
  .names = "{.col}_ctr"
)
}
\arguments{
\item{frame}{Dataframe in which variables are found}

\item{choose}{Which variables should be centered?}

\item{center}{Center variables? Default is TRUE}

\item{scale}{Scale variables? Default is FALSE}

\item{...}{Other arguments passed into \code{dplyr::mutate}, especially useful to
pass \code{.by}}

\item{.names}{How should new variables be named, default is to append "_ctr"}
}
\value{
Returns same dataframe but with additionally calculated variables.
}
\description{
Function is a wrapper around a \code{dplyr::mutate} and \code{dplyr::across}
combination which can centre and/or standardize the specified variables.
Allows for group specification through \code{.by = X}.
}
