% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Clean string vector}
\usage{
clean_colnames(
  object,
  sep = ".",
  names = TRUE,
  numlead = FALSE,
  case = c("title", "lower", "upper", "sentence")
)
}
\arguments{
\item{object}{Vector of column names or named vector. If a dataframe is
provided, also specify \code{names = TRUE}, this is implemented to prevent
confusion}

\item{sep}{Value to use to separate alphanumeric characters, default is a
period "."}

\item{names}{Clean names of an object? Default is \code{FALSE}, use \code{TRUE} for
objects where the names should be cleaned rather than the values}

\item{numlead}{Should numerical values be allowed at the start of a string?
Default is \code{FALSE}}

\item{case}{What case should values be returned in? Default is unchanged,
cases such as Title Case, lower case, UPPER CASE, or Sentence case
supported. Case is implemented before separation, so the only difference
between title and sentence is if input includes spaces to separate words.}
}
\description{
Cleans up a given string vector, particularly useful if you have a vector of
names you want to work with (e.g., converting lowercase to Sentence case or
replacing underscore separators with period separators). Can also be extended
to data frames either through providing a dataframe object with \code{names = TRUE} or through \code{clean_colnames}. Depends on \code{stringr}.
}
