% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{clean_string}
\alias{clean_string}
\title{Cleans a provided string}
\usage{
clean_string(
  object,
  sep = ".",
  names = FALSE,
  numlead = FALSE,
  case = c("title", "lower", "upper", "sentence")
)
}
\arguments{
\item{object}{Input, either vector or named object (specify \code{names = TRUE})}

\item{sep}{What character(s) should be used to separate words within cells}

\item{names}{Specify whether the cleaning should be done on the names of a
provided object rather than the object itself}

\item{numlead}{Allows for numbers to remain as the lead character in a
string, default is \code{FALSE}}

\item{case}{What case should be returned? Allows for title, lower, upper, and
sentence}
}
\value{
Vector or object as the same type received, but with text changes.
}
\description{
Cleans a provided string with specified parameters, used primarily as a
helper for cleaning column names.
}
