% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{corrplot}
\alias{corrplot}
\title{Plot correlations}
\usage{
corrplot(
  frame,
  groups = NULL,
  by = NULL,
  type = "pearson",
  alpha = 0.05,
  p.adjust = "none",
  digits = 2,
  auto.name = FALSE,
  colors = list(low = "red", med = "white", high = "blue")
)
}
\arguments{
\item{frame}{Dataframe to find correlations from, must be numeric except for grouping variable}

\item{groups}{Groups to use with a format list(groups = c("y", "x")); groups will be placed along the y triangle or x triangle based on placement}

\item{by}{Variables to sort plot by, limits x axis. Behavior with groups will likely not provide the desired output}

\item{type}{Type of correlations, Pearson or Spearman}

\item{alpha}{Alpha level, default is .05. Correlations significant below alpha will be colored, those above alpha will not be colored}

\item{p.adjust}{Should p values be adjusted before plotting?}

\item{digits}{Digits to display in the plot, default is 2}

\item{auto.name}{Should names be automatically capitalized and separated by spaces? Default is FALSE}

\item{colors}{How should values be colored? Accepts a list of color specifications with default colors = list(low = "red", med = "white", high = "blue")}
}
\value{
ggplot object
}
\description{
Creates a plot of correlations for one group or two.
}
