% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{corrsBy}
\alias{corrsBy}
\title{Correlations by group or specified list of variables}
\usage{
corrsBy(
  frame,
  specs = NULL,
  groups = NULL,
  by = NULL,
  type = "pearson",
  alpha = Inf,
  p.adjust = "none",
  digits = 3,
  autocor = FALSE
)
}
\arguments{
\item{frame}{Matrix or dataframe of variables to correlate. Requires numeric except for grouping variable(s)}

\item{groups}{Retained for backwards compatibility. Vector of columns to group by, can be several using c()}

\item{by}{Retained for backwards compatibility. Vector of columns to display correlations by. Inspired from function in SPSS. Provides correlations y ~ x, where y is every other variable in the dataframe and x are the variable names provided for this argument. Can accept formula in the form . ~ x1 + x2 + [...].}

\item{type}{Pearson or spearman correlations}

\item{alpha}{Alpha level to display correlations, defaults to Inf (all printed)}

\item{p.adjust}{P level adjustment using p.adjust.methods. Evaluated after all other steps except rounding. Under the hood, all possible correlations are conducted (full correlation matrix) however, adjustments are only done on those of interest (e.g., those implied via by argument)}

\item{digits}{Rounding for displayed results, defaults to 3}

\item{autocor}{Should self-correlations (always 1) be included/ displayed, default is FALSE}
}
\description{
Convenience function which allows correlations to be found by multiple groups or based on a set of specified focal variables (specified in by). Can conduct pearson or spearman correlations and filter output by specified alpha level.
}
\examples{

test_frame <- as.data.frame(
  matrix(rnorm(600), ncol = 10)
)
test_frame$group1 <- rep(c("A", "B"), 30)
test_frame$group2 <- rep(c("X", "Y", "Z"), 20)

}
