% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_model.R
\name{describe_model}
\alias{describe_model}
\title{Describe model}
\usage{
describe_model(formula, data, model = NULL, digits = 2)
}
\arguments{
\item{formula}{Formula specification in the form of y ~ x. Note that lme4 syntax (y ~ x + (1 | j)) and correlation format ( . ~ x + y) are not supported}

\item{data}{Dataframe to reference}

\item{model}{Model object to provide descriptives for, will determine formula and data from the provided model object and flag if model type is not currently supported}

\item{digits}{Rounding digit, default is 2}
}
\value{
List of dataframes with descriptives by each term.
}
\description{
Uses a formula interface to provide descriptive statistics per model term. Provides mean of outcome, sd of outcome, Pearson's r for continuous predictors with outcome; all based on categorical variables. Formula interface is for ability to translate model specifications into descriptives, future functionality will add ability to specify a model instead.
}
