% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_model.R
\name{describe_term}
\alias{describe_term}
\title{Descriptives for a given term in a model}
\usage{
describe_term(term, data, y_term, digits = 2)
}
\description{
Helper function for model_desc which is looped
}
