% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm_effectsize.R
\name{emm_cohend}
\alias{emm_cohend}
\title{Cohen's d using emmeans}
\usage{
emm_cohend(
  model,
  formula,
  within,
  standardize,
  level = 0.95,
  var = NULL,
  verbose = FALSE,
  digits = 4,
  return = c("print", "frame"),
  at = NULL,
  ...
)
}
\arguments{
\item{model}{Model to identify emmeans, supports most models supported by emmeans; however, your mileage may vary. In particular, built in support for "lm", "lme", "glm", "lmerTest", "lme4", "lmerModLmerTest", "glmerMod", and "glmmTMB"}

\item{formula}{emmeans formula with the format of contrast ~ x | w}

\item{within}{Used to calculated error for Cohen's d, formula of format x ~ b + (k) where x is the within-group variable, b is a between-group (optional), and k defines within-group clusters with the format (level2/level1). Cohen's d calculated is dav (Lakens, 2013) when within is specified}

\item{level}{Specifies level for confidence intervals, default is 95%}

\item{var}{Supports use of emtrends and finding Cohen's d for difference in simple slopes}

\item{digits}{Digits to round to, default is 4}

\item{return}{How to return output, default is to print}

\item{...}{Additional arguments to pass to emmeans, including type = "response" to get odds ratios. Behavior may be unexpected with certain other arguments, such as at, or if a different type is specified}
}
\description{
Convenience function which wraps around emmeans to provide Cohen's d and confidence intervals. Also rounds and can output a dataframe, for convenience. Printed output rounds to specified digits, frame output does not. Supports repeated-measures or within-group calculation of CI using within. Note that this function requires contrasts, it will fail otherwise.
}
