% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm_effectsize.R
\name{emm_oddratio}
\alias{emm_oddratio}
\title{Odds ratio using emmeans}
\usage{
emm_oddratio(
  model,
  formula,
  level = 0.95,
  standardize,
  return = c("print", "frame"),
  digits = 4,
  ...
)
}
\arguments{
\item{model}{Model to identify emmeans, supports any model supported by emmeans}

\item{formula}{emmeans formula with the format of contrast ~ x | w}

\item{level}{Specifies level for confidence intervals, default is 95%}

\item{return}{How to return output, default is to print}

\item{digits}{Digits to round to, default is 4}

\item{...}{Additional arguments to pass to emmeans, including type = "response" to get odds ratios. Behavior may be unexpected with certain other arguments, such as at, or if a different type is specified}
}
\description{
Convenience wrapper around emmeans for calculating odds ratios. This functionality is well covered by emmeans already, all this adds is the ability to specify confidence intervals for the odds ratio. Also rounds and can output a dataframe, for convenience. Printed output rounds to specified digits, frame output does not. Note that this function requires contrasts, it will fail otherwise
}
