% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_k.R
\name{get_k}
\alias{get_k}
\title{Get K}
\usage{
get_k(frame, n.var = NULL, k.var, digits = 2)
}
\arguments{
\item{frame}{Dataframe to provide n and k for}

\item{n.var}{Variable identifying grouping for n, can be left blank}

\item{k.var}{Variable identifying level of k, e.g., id}

\item{digits}{Digits for rounding, default is 2}
}
\description{
Function to receive a dataframe and output the n for a grouping variable and k for each observation within that group. E.g., for EMA data where you have k observations per n individuals in multiple groups.
}
