% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode of a vector}
\usage{
mode(x)
}
\arguments{
\item{x}{Vector to return mean of}
}
\description{
Returns the mode of a given vector, will convert factors and dates to characters!
}
\examples{

x <- round(rnorm(100), 0)
x <- rnorm(100)

}
