% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_labels.R
\name{sig_labels}
\alias{sig_labels}
\title{Significance labels}
\usage{
sig_labels(vector, trend = FALSE, fourstar = FALSE)
}
\arguments{
\item{vector}{Vector of *p* values}

\item{trend}{Should trend daggers (\U2020) be displayed? Default is FALSE}

\item{fourstar}{Should values < .0001 be labeled with four stars? Default is FALSE}
}
\value{
Vector of labels of same length as input.
}
\description{
Converts a vector of *p* values to APA-style significance labels. Default performance is ` ` = *p* > .05, `*` = *p* < .05, `**` = *p* < .01, `***` = *p* < .001.
}
\examples{
 p.values <- runif(100, min = 0, max = .5)
 stars <- sig_labels(p.values, trend = T)

}
