% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_frame.R
\name{standardize_frame}
\alias{standardize_frame}
\title{Standardize a dataframe}
\usage{
standardize_frame(
  frame,
  center = TRUE,
  scale = TRUE,
  skip = NULL,
  choose = NULL,
  scale.factors = FALSE,
  scale.dichotomous = FALSE,
  grouping = NULL
)
}
\arguments{
\item{frame}{Dataframe to standardize values in}

\item{center}{Should data be centered? Default is TRUE. Will center on grand mean by default, grouping can be used to specify groups for group-mean centering}

\item{scale}{Should data be scaled? Default is TRUE}

\item{skip}{Which columns should be skipped? Default is any non-numeric and any column where only two values are observed (e.g., 0, 1)}

\item{choose}{Which columns should be scaled? Useful if some numeric variables should be centered while others shouldn't. Complement to skip}

\item{scale.factors}{Should factors be converted to numeric and scaled? Default is FALSE}

\item{scale.dichotomous}{Should dichotomous variables be scaled? Default is FALSE}

\item{grouping}{Should dataframe be grouped prior to scaling? Useful for group-mean or idiographic scalings}
}
\description{
Function receives an input dataframe and outputs a standardized version, with some additional helpers on what to standardize and how.
}
