% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_model}
\alias{standardize_model}
\title{Standardize model}
\usage{
standardize_model(model, ...)
}
\arguments{
\item{model}{Model to standardize data for}

\item{...}{Arguments to pass to \code{standardize_frame}}
}
\value{
Model with the same specifications and formula but standardized
continuous variables
}
\description{
Designed as a shortcut for standardizing (z-scoring) the data used for a
given model. The \code{parameters} package allows for standardized parameters,
this is designed to work with other functions that work on a model rather
than the parameters output, such as \code{emmeans}.
}
